# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Polls API",
        "description": "API for creating polls, voting, and fetching results",
        "license": {
            "name": "BSD License"
        },
        "version": "v1"
    },
    "basePath": "/api",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT or Token authorization. Example: \"Bearer {token}\""
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
    "paths": {
...
        "/polls/": {
            "get": {
                "operationId": "polls_list",
                "description": "Retrieve a list of pollswith filtering options",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "A search term.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ordering",
                        "in": "query",
                        "description": "Which field to use when ordering the results.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "A page number within the paginated result set.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "question",
                        "in": "query",
                        "description": "Filter by question text (contains)",
                        "type": "string"
                    },
                    {
                        "name": "creator_email",
                        "in": "query",
                        "description": "Filter by creator_email (contains)",
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (active, upcoming, expired)",
                        "type": "string",
                        "enum": [
                            "active",
                            "upcoming",
                            "expired"
                        ]
                    },
                    {
                        "name": "created_after",
                        "in": "query",
                        "description": "Filter polls created after this date",
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "created_before",
                        "in": "query",
                        "description": "Filter polls created before this date",
                        "type": "string",
                        "format": "date-time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "required": [
                                "count",
                                "results"
                            ],
                            "type": "object",
                            "properties": {
                                "count": {
                                    "type": "integer"
                                },
                                "next": {
                                    "type": "string",
                                    "format": "uri",
                                    "x-nullable": true
                                },
                                "previous": {
                                    "type": "string",
                                    "format": "uri",
                                    "x-nullable": true
                                },
                                "results": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Poll"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "polls"
                ]
            },
            "post": {
                "operationId": "polls_create",
                "description": "Create a new poll",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PollCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Poll"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "tags": [
                    "polls"
                ]
            },
            "parameters": []
        },
...
    },
    "definitions": {
...
        "Poll": {
            "required": [
                "question",
                "options"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "Id",
                    "type": "string",
                    "format": "uuid",
                    "readOnly": true
                },
                "question": {
                    "title": "Question",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "options": {
                    "title": "Options",
                    "type": "object"
                },
                "is_anonymous": {
                    "title": "Is anonymous",
                    "type": "boolean"
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "title": "Updated at",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "owner_email": {
                    "title": "Owner email",
                    "type": "string",
                    "readOnly": true
                },
                "creator_email": {
                    "title": "Creator email",
                    "type": "string",
                    "readOnly": true
                },
                "start_date": {
                    "title": "Start date",
                    "type": "string",
                    "format": "date-time"
                },
                "expiry_date": {
                    "title": "Expiry date",
                    "type": "string",
                    "format": "date-time"
                },
                "is_active": {
                    "title": "Is active",
                    "type": "boolean"
                },
                "total_votes": {
                    "title": "Total votes",
                    "type": "string",
                    "readOnly": true
                },
                "has_user_voted": {
                    "title": "Has user voted",
                    "type": "string",
                    "readOnly": true
                },
                "status": {
                    "title": "Status",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "PollCreate": {
            "required": [
                "question",
                "options"
            ],
            "type": "object",
            "properties": {
                "question": {
                    "title": "Question",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "options": {
                    "title": "Options",
                    "type": "object"
                },
                "is_anonymous": {
                    "title": "Is anonymous",
                    "type": "boolean"
                },
                "start_date": {
                    "title": "Start date",
                    "type": "string",
                    "format": "date-time"
                },
                "expiry_date": {
                    "title": "Expiry date",
                    "type": "string",
                    "format": "date-time"
                },
                "owner_email": {
                    "title": "Owner email",
                    "type": "string",
                    "readOnly": true
                },
                "creator_email": {
                    "title": "Creator email",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
...
    }
}